doc: initial analysis of word distribution - histogram
Begin a simple analysis of the English words distribution using
LibreOffice .ods spreadsheet, to just quickly visualize the
distribution.
Start with just a simple histogram.

Now we have the same data duplicated in two files, which is a little
tedious. Would be nice to de-duplicate that, have a single source of
truth.